+++ src/index.ts
@@ -1,5 +1,5 @@
-export * as calendar from './lib/api/calendar';
-export * as youtube from './lib/api/youtube';
+export { getAPI as getCalendarAPI } from './lib/api/calendar';
+export { getAPI as getYoutubeAPI } from './lib/api/youtube';
 export { getItems } from './lib/api/shared';
 export { createProfile, getProfiles } from './lib/profiles';
 export { login, getAuth } from './lib/auth';