+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0](../../tags/v5.0.0) - 2023-03-20
+### Added
+- Non-persistence mode for getAuth: ability to not save sensitive credentials into the file
+
+## Removed
+- Removed getter methods from api helpers to be able to re-use auth between api usages. Now better call `getAPI` and then run needed native methods on their own.
+
 ## [4.0.0](../../tags/v4.0.0) - 2023-03-12
 ### Changed
 - Silent mode for getting items